@@ -10,17 +10,19 @@
     "eject": "react-scripts-ts eject"
   },
   "dependencies": {
-    "@material-ui/core": "^3.9.2",
-    "@material-ui/icons": "^3.0.2",
+    "@date-io/date-fns": "^1.3.9",
+    "@material-ui/core": "^4.4.0",
+    "@material-ui/icons": "^4.2.1",
+    "@material-ui/pickers": "^3.2.4",
     "classnames": "^2.2.6",
-    "material-ui-slider": "^3.0.0",
+    "date-fns": "^2.0.1",
     "openlayers": "^4.6.5",
-    "react": "^16.8.0",
-    "react-dom": "^16.8.0",
+    "react": "^16.9.0",
+    "react-dom": "^16.9.0",
     "react-redux": "^6.0.0",
     "react-scripts-ts": "3.1.0",
-    "recharts": "^1.4.2",
-    "redux": "^4.0.1",
+    "recharts": "^1.7.1",
+    "redux": "^4.0.4",
     "redux-logger": "^3.0.6",
     "redux-thunk": "^2.3.0",
     "reselect": "^4.0.0"
@@ -28,19 +30,23 @@
   "devDependencies": {
     "@types/classnames": "^2.2.7",
     "@types/enzyme": "^3.1.17",
-    "@types/enzyme-adapter-react-16": "^1.0.3",
+    "@types/enzyme-adapter-react-16": "^1.0.5",
     "@types/geojson": "^7946.0.5",
     "@types/jest": "^24.0.0",
     "@types/node": "^10.12.21",
     "@types/openlayers": "^4.6.15",
-    "@types/react": "^16.8.1",
+    "@types/react": "^16.9.2",
     "@types/react-dom": "^16.0.11",
-    "@types/react-redux": "^7.0.1",
-    "@types/recharts": "^1.1.10",
-    "@types/redux-logger": "^3.0.6",
+    "@types/react-redux": "^7.1.2",
+    "@types/recharts": "=1.1.10",
+    "@types/redux-logger": "^3.0.7",
     "@types/redux-thunk": "^2.1.0",
     "enzyme": "^3.8.0",
     "enzyme-adapter-react-16": "^1.9.1",
-    "typescript": "^3.3.1"
+    "typescript": "^3.6.2"
+  },
+  "resolutions": {
+    "react": "^16.9.0",
+    "@types/react": "^16.9.2"
   }
 }
