@@ -1,21 +1,25 @@
 import * as React from 'react';
 import { createStyles, withStyles, WithStyles } from '@material-ui/core/styles';
 import { Theme } from '@material-ui/core';
-
-const Slider = require('material-ui-slider').Slider;
+import Box from "@material-ui/core/Box";
+import Slider, { Mark } from "@material-ui/core/Slider";
 
 import { Time, TimeRange, UNIT } from '../model/timeSeries';
-import TimeRangeContainer from "./TimeRangeContainer";
-import FormControl from "@material-ui/core/FormControl";
+import {
+    utcTimeToLocalDateTimeString,
+    utcTimeToLocalIsoDateString,
+} from "../util/time";
 
+const HOR_MARGIN = 5;
 
 // noinspection JSUnusedLocalSymbols
 const styles = (theme: Theme) => createStyles(
     {
-        formControl: {
-            marginRight: theme.spacing.unit * 2,
-            marginBottom: theme.spacing.unit,
-            minWidth: 400,
+        box: {
+            marginTop: theme.spacing(2),
+            marginLeft: theme.spacing(HOR_MARGIN),
+            marginRight: theme.spacing(HOR_MARGIN),
+            minWidth: 200,
         },
     }
 );
@@ -25,7 +29,6 @@ interface TimeSliderProps extends WithStyles<typeof styles> {
     selectTime?: (time: Time | null) => void;
     selectedTimeRange?: TimeRange | null;
     selectTimeRange?: (timeRange: TimeRange | null) => void;
-    visibleTimeRange?: TimeRange | null;
 }
 
 interface TimeSliderState {
@@ -37,63 +40,61 @@ class TimeSlider extends React.Component<TimeSliderProps, TimeSliderState> {
 
     constructor(props: TimeSliderProps) {
         super(props);
-        this.state = {selectedTime: this.props.selectedTime};
+        this.state = {
+            selectedTime: props.selectedTime
+        };
     }
 
-    componentDidUpdate(prevProps: Readonly<TimeSliderProps>, prevState: Readonly<TimeSliderState>): void {
-        let selectedTime = this.props.selectedTime;
-        if (selectedTime !== prevProps.selectedTime) {
-            this.setState({selectedTime});
-        }
+    // noinspection JSUnusedGlobalSymbols
+    static getDerivedStateFromProps(nextProps: Readonly<TimeSliderProps>) {
+        return {selectedTime: nextProps.selectedTime};
     }
 
-    handleSliderChange = (selectedTime: Time) => {
-        this.setState({selectedTime});
-    };
-
-    handleChangeComplete = (selectedTime: Time) => {
-        if (this.props.selectTime) {
-            this.props.selectTime(selectedTime);
+    handleChange = (event: React.ChangeEvent<{}>, value: number | number[]) => {
+        if (typeof value === 'number') {
+            this.setState({selectedTime: value});
         }
     };
 
-    handleSelectedTimeRangeChange = (selectedTimeRange: TimeRange) => {
-        if (this.props.selectTimeRange) {
-            this.props.selectTimeRange([selectedTimeRange[0], selectedTimeRange[1]]);
+    handleChangeCommitted = (event: React.ChangeEvent<{}>, value: number | number[]) => {
+        if (this.props.selectTime && typeof value === 'number') {
+            this.props.selectTime(value as number);
         }
     };
 
     render() {
-        let {classes, visibleTimeRange, selectedTimeRange, selectedTime} = this.props;
+        let {classes, selectedTimeRange} = this.props;
 
-        let visibleDataRange = Array.isArray(visibleTimeRange);
-        if (!visibleDataRange) {
-            visibleTimeRange = [Date.now() - 2 * UNIT.years, Date.now()];
+        let selectedTimeRangeValid = Array.isArray(selectedTimeRange);
+        if (!selectedTimeRangeValid) {
+            selectedTimeRange = [Date.now() - 2 * UNIT.years, Date.now()];
         }
-        if (!Array.isArray(selectedTimeRange)) {
-            selectedTimeRange = visibleTimeRange;
+
+        const selectedTime = this.state.selectedTime || selectedTimeRange![0];
+
+        const marks: Mark[] = [
+            {value: selectedTimeRange![0], label: utcTimeToLocalIsoDateString(selectedTimeRange![0])},
+            {value: selectedTimeRange![1], label: utcTimeToLocalIsoDateString(selectedTimeRange![1])},
+        ];
+
+        function valueLabelFormat(value: number) {
+            return utcTimeToLocalDateTimeString(value);
         }
 
         return (
-            <FormControl className={classes.formControl}>
-                <TimeRangeContainer
-                    min={visibleTimeRange![0]}
-                    max={visibleTimeRange![1]}
-                    step={UNIT.weeks}
-                    value={selectedTimeRange || null}
-                    onChange={this.handleSelectedTimeRangeChange}
-                    disabled={!visibleDataRange}
-                >
-                    <Slider
-                        disabled={!visibleDataRange}
-                        min={selectedTimeRange![0]}
-                        max={selectedTimeRange![1]}
-                        value={selectedTime}
-                        onChange={this.handleSliderChange}
-                        onChangeComplete={this.handleChangeComplete}
-                    />
-                </TimeRangeContainer>
-            </FormControl>
+            <Box className={classes.box}>
+                <Slider
+                    disabled={!selectedTimeRangeValid}
+                    min={selectedTimeRange![0]}
+                    max={selectedTimeRange![1]}
+                    value={selectedTime}
+                    valueLabelDisplay="off"
+                    valueLabelFormat={valueLabelFormat}
+                    marks={marks}
+                    onChange={this.handleChange}
+                    onChangeCommitted={this.handleChangeCommitted}
+                />
+            </Box>
         );
     }
 }
