@@ -1,29 +1,22 @@
 import { Dispatch } from 'redux';
-import * as geojson from 'geojson'
-import * as ol from 'openlayers';
 
 import {
-    selectedDatasetIdSelector,
     selectedDatasetSelectedPlaceGroupsSelector,
     selectedDatasetSelector,
-    selectedDatasetTimeDimensionSelector,
-    selectedDatasetVariableSelector,
     selectedServerSelector,
 } from '../selectors/controlSelectors';
 import { Dataset } from '../model/dataset';
-import { Time, TimeRange, TimeSeries } from '../model/timeSeries';
+import { Time, TimeRange } from '../model/timeSeries';
 import { AppState } from '../states/appState';
 import * as api from '../api'
 import { MessageLogAction, postMessage } from './messageLogActions';
 import {
     updateDatasetPlaceGroup, UPDATE_DATASET_PLACE_GROUP,
     UpdateDatasetPlaceGroup,
-    UpdateTimeSeries,
-    updateTimeSeries
 } from './dataActions';
-import { MAP_OBJECTS } from '../states/controlState';
 import { isValidPlaceGroup, PlaceGroup } from '../model/place';
 import { I18N } from '../config';
+import { ControlState, TimeAnimationInterval } from "../states/controlState";
 
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
@@ -53,16 +46,14 @@ export type SELECT_PLACE_GROUPS = typeof SELECT_PLACE_GROUPS;
 export interface SelectPlaceGroups {
     type: SELECT_PLACE_GROUPS;
     selectedPlaceGroupIds: string[] | null;
-    // TODO: Having datasets in here is ugly, but we need it in the reducer.
-    datasets: Dataset[];
 }
 
-export function selectPlaceGroups(selectedPlaceGroupIds: string[] | null, datasets: Dataset[]) {
+export function selectPlaceGroups(selectedPlaceGroupIds: string[] | null) {
 
     return (dispatch: Dispatch<SelectPlaceGroups | UpdateDatasetPlaceGroup | AddActivity | RemoveActivity | MessageLogAction>, getState: () => AppState) => {
         const apiServer = selectedServerSelector(getState());
 
-        dispatch(_selectPlaceGroups(selectedPlaceGroupIds, datasets));
+        dispatch(_selectPlaceGroups(selectedPlaceGroupIds));
 
         const dataset = selectedDatasetSelector(getState());
         const placeGroups = selectedDatasetSelectedPlaceGroupsSelector(getState());
@@ -89,8 +80,8 @@ export function selectPlaceGroups(selectedPlaceGroupIds: string[] | null, datase
     };
 }
 
-export function _selectPlaceGroups(selectedPlaceGroupIds: string[] | null, datasets: Dataset[]): SelectPlaceGroups {
-    return {type: SELECT_PLACE_GROUPS, selectedPlaceGroupIds, datasets};
+export function _selectPlaceGroups(selectedPlaceGroupIds: string[] | null): SelectPlaceGroups {
+    return {type: SELECT_PLACE_GROUPS, selectedPlaceGroupIds};
 }
 
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -103,10 +94,12 @@ export interface SelectPlace {
     selectedPlaceId: string | null;
     // TODO: Having datasets in here is ugly, but we need it in the reducer.
     datasets: Dataset[];
+    // TODO: Having userPlaceGroup in here is ugly, but we need it in the reducer.
+    userPlaceGroup: PlaceGroup;
 }
 
-export function selectPlace(selectedPlaceId: string | null, datasets: Dataset[]): SelectPlace {
-    return {type: SELECT_PLACE, selectedPlaceId, datasets};
+export function selectPlace(selectedPlaceId: string | null, datasets: Dataset[], userPlaceGroup: PlaceGroup): SelectPlace {
+    return {type: SELECT_PLACE, selectedPlaceId, datasets, userPlaceGroup};
 }
 
 
@@ -122,21 +115,6 @@ export interface SelectVariable {
 
 export function selectVariable(selectedVariableName: string | null): SelectVariable {
     return {type: SELECT_VARIABLE, selectedVariableName};
-}
-
-//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-export const SELECT_USER_PLACE = 'SELECT_USER_PLACE';
-export type SELECT_USER_PLACE = typeof SELECT_USER_PLACE;
-
-export interface SelectUserPlace {
-    type: SELECT_USER_PLACE;
-    selectedUserPlaceId: string | null;
-}
-
-// TODO (forman): let users select their points
-export function selectUserPlace(selectedUserPlaceId: string | null): SelectUserPlace {
-    return {type: SELECT_USER_PLACE, selectedUserPlaceId};
 }
 
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -183,20 +161,6 @@ export function selectTimeRange(selectedTimeRange: TimeRange | null): SelectTime
 
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-export const UPDATE_VISIBLE_TIME_RANGE = 'UPDATE_VISIBLE_TIME_RANGE';
-export type UPDATE_VISIBLE_TIME_RANGE = typeof UPDATE_VISIBLE_TIME_RANGE;
-
-export interface UpdateVisibleTimeRange {
-    type: UPDATE_VISIBLE_TIME_RANGE;
-    visibleTimeRange: TimeRange | null;
-}
-
-export function updateVisibleTimeRange(visibleTimeRange: TimeRange | null): UpdateVisibleTimeRange {
-    return {type: UPDATE_VISIBLE_TIME_RANGE, visibleTimeRange};
-}
-
-//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
 export const SELECT_TIME_SERIES_UPDATE_MODE = 'SELECT_TIME_SERIES_UPDATE_MODE';
 export type SELECT_TIME_SERIES_UPDATE_MODE = typeof SELECT_TIME_SERIES_UPDATE_MODE;
 
@@ -217,109 +181,11 @@ export type UPDATE_TIME_ANIMATION = typeof UPDATE_TIME_ANIMATION;
 export interface UpdateTimeAnimation {
     type: UPDATE_TIME_ANIMATION;
     timeAnimationActive: boolean;
-    timeAnimationInterval: number;
+    timeAnimationInterval: TimeAnimationInterval;
 }
 
-export function updateTimeAnimation(timeAnimationActive: boolean, timeAnimationInterval: number): UpdateTimeAnimation {
+export function updateTimeAnimation(timeAnimationActive: boolean, timeAnimationInterval: TimeAnimationInterval): UpdateTimeAnimation {
     return {type: UPDATE_TIME_ANIMATION, timeAnimationActive, timeAnimationInterval};
-}
-
-//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-export const ADD_GEOMETRY = 'ADD_GEOMETRY';
-export type ADD_GEOMETRY = typeof ADD_GEOMETRY;
-
-export interface AddGeometry {
-    type: ADD_GEOMETRY;
-    featureId: string;
-    geometry: geojson.Geometry;
-}
-
-export function addGeometry(featureId: string, geometry: geojson.Geometry, color: string) {
-    return (dispatch: Dispatch<AddGeometry | UpdateTimeSeries | MessageLogAction>, getState: () => AppState) => {
-        const apiServer = selectedServerSelector(getState());
-
-        dispatch(_addGeometry(featureId, geometry));
-
-        // TODO (forman): extract following code as separate action addTimeSeries()
-
-        const selectedDatasetId = selectedDatasetIdSelector(getState());
-        const selectedDatasetTimeDim = selectedDatasetTimeDimensionSelector(getState());
-        const selectedVariable = selectedDatasetVariableSelector(getState());
-        const timeSeriesUpdateMode = getState().controlState.timeSeriesUpdateMode;
-
-        const timeLabels = selectedDatasetTimeDim != null ? selectedDatasetTimeDim.labels : null;
-
-        if (selectedDatasetId && selectedVariable && timeLabels !== null) {
-            const numTimeLabels = timeLabels.length;
-            const timeChunkSize = 16;
-            let endTimeIndex = numTimeLabels - 1;
-            let startTimeIndex = endTimeIndex - timeChunkSize + 1;
-
-            const getTimeSeriesChunk = () => {
-                const startDateLabel = startTimeIndex >= 0 ? timeLabels[startTimeIndex] : null;
-                const endDateLabel = timeLabels[endTimeIndex];
-                return api.getTimeSeriesForGeometry(apiServer.url,
-                                                    selectedDatasetId,
-                                                    selectedVariable,
-                                                    featureId,
-                                                    geometry,
-                                                    startDateLabel,
-                                                    endDateLabel);
-            };
-
-            const successAction = (timeSeries: TimeSeries | null) => {
-                if (timeSeries !== null && hasUserFeature(featureId)) {
-                    const hasMore = startTimeIndex > 0;
-                    const dataProgress = hasMore ? (numTimeLabels - startTimeIndex) / numTimeLabels : 1.0;
-                    dispatch(updateTimeSeries({...timeSeries, dataProgress, color},
-                                              timeSeriesUpdateMode,
-                                              endTimeIndex === numTimeLabels - 1 ? 'new' : 'append'));
-                    if (hasMore && hasUserFeature(featureId)) {
-                        startTimeIndex -= timeChunkSize;
-                        endTimeIndex -= timeChunkSize;
-                        getTimeSeriesChunk().then(successAction);
-                    }
-                } else {
-                    dispatch(postMessage('info', 'No data found here'));
-                }
-            };
-
-            getTimeSeriesChunk()
-                .then(successAction)
-                .catch((error: any) => {
-                    dispatch(postMessage('error', error + ''));
-                });
-        }
-    };
-}
-
-function _addGeometry(featureId: string, geometry: geojson.Geometry): AddGeometry {
-    return {type: ADD_GEOMETRY, featureId, geometry};
-}
-
-function hasUserFeature(featureId: string): boolean {
-    if (MAP_OBJECTS.userLayer) {
-        const userLayer = MAP_OBJECTS.userLayer as ol.layer.Vector;
-        const source = userLayer.getSource();
-        return !!source.getFeatureById(featureId);
-    }
-    return false;
-}
-
-//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-export const SELECT_GEOMETRY = 'SELECT_GEOMETRY';
-export type SELECT_GEOMETRY = typeof SELECT_GEOMETRY;
-
-export interface SelectGeometry {
-    type: SELECT_GEOMETRY;
-    featureId: string;
-    geometry: geojson.Geometry;
-}
-
-export function selectGeometry(featureId: string, geometry: geojson.Geometry): SelectGeometry {
-    return {type: SELECT_GEOMETRY, featureId, geometry};
 }
 
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -395,23 +261,34 @@ export function closeDialog(dialogId: string): CloseDialog {
 
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
+export const UPDATE_SETTINGS = 'UPDATE_SETTINGS';
+export type UPDATE_SETTINGS = typeof UPDATE_SETTINGS;
+
+export interface UpdateSettings {
+    type: UPDATE_SETTINGS;
+    settings: ControlState;
+}
+
+export function updateSettings(settings: ControlState): UpdateSettings {
+    return {type: UPDATE_SETTINGS, settings};
+}
+
+//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
 export type ControlAction =
     SelectDataset
     | UpdateDatasetPlaceGroup
     | SelectVariable
     | SelectPlaceGroups
     | SelectPlace
-    | SelectUserPlace
     | SelectTime
     | IncSelectedTime
     | SelectTimeRange
-    | AddGeometry
-    | SelectGeometry
     | SelectTimeSeriesUpdateMode
-    | UpdateVisibleTimeRange
     | UpdateTimeAnimation
     | AddActivity
     | RemoveActivity
     | ChangeLocale
+    | UpdateSettings
     | OpenDialog
     | CloseDialog;
