@@ -3,23 +3,22 @@ import { withStyles, WithStyles, createStyles, Theme } from '@material-ui/core/s
 import Input from '@material-ui/core/Input';
 import InputLabel from '@material-ui/core/InputLabel';
 import MenuItem from '@material-ui/core/MenuItem';
-import FormControl from '@material-ui/core/FormControl';
 import Select from '@material-ui/core/Select';
 
 import { Dataset } from '../model/dataset';
 import { WithLocale } from '../util/lang';
 import { I18N } from '../config';
+import ControlBarItem from './ControlBarItem';
 
 
 const styles = (theme: Theme) => createStyles(
     {
         formControl: {
-            marginRight: theme.spacing.unit * 2,
-            marginBottom: theme.spacing.unit,
+            marginRight: theme.spacing(2),
+            marginBottom: theme.spacing(1),
             minWidth: 120,
         },
         selectEmpty: {
-            marginTop: theme.spacing.unit * 2,
         },
     });
 
@@ -41,30 +40,38 @@ class DatasetSelect extends React.Component<DatasetSelectProps> {
         const selectedDatasetId = this.props.selectedDatasetId || '';
         const datasets = this.props.datasets || [];
 
+        const datasetSelectLabel = (
+            <InputLabel shrink htmlFor="dataset-select">
+            {I18N.get('Dataset')}
+        </InputLabel>
+        );
+
+        const datasetSelect = (
+            <Select
+                value={selectedDatasetId}
+                onChange={this.handleDatasetChange}
+                input={<Input name="dataset" id="dataset-select"/>}
+                displayEmpty
+                name="dataset"
+                className={classes.selectEmpty}
+            >
+                {datasets.map(dataset => (
+                    <MenuItem
+                        key={dataset.id}
+                        value={dataset.id}
+                        selected={dataset.id === selectedDatasetId}
+                    >
+                        {dataset.title}
+                    </MenuItem>
+                ))}
+            </Select>
+        );
+
         return (
-            <FormControl className={classes.formControl}>
-                <InputLabel shrink htmlFor="dataset-select">
-                    {I18N.get("Dataset")}
-                </InputLabel>
-                <Select
-                    value={selectedDatasetId}
-                    onChange={this.handleDatasetChange}
-                    input={<Input name="dataset" id="dataset-select"/>}
-                    displayEmpty
-                    name="dataset"
-                    className={classes.selectEmpty}
-                >
-                    {datasets.map(dataset => (
-                        <MenuItem
-                            key={dataset.id}
-                            value={dataset.id}
-                            selected={dataset.id === selectedDatasetId}
-                        >
-                            {dataset.title}
-                        </MenuItem>
-                    ))}
-                </Select>
-            </FormControl>
+            <ControlBarItem
+                label={datasetSelectLabel}
+                control={datasetSelect}
+            />
         );
     }
 }
