@@ -28,22 +28,22 @@ import { newId } from '../util/id';
 const styles = (theme: Theme) => createStyles(
     {
         formControl: {
-            marginLeft: theme.spacing.unit,
-            marginRight: theme.spacing.unit,
+            marginLeft: theme.spacing(1),
+            marginRight: theme.spacing(1),
             width: 200,
         },
         textField: {
-            marginLeft: theme.spacing.unit,
-            marginRight: theme.spacing.unit,
+            marginLeft: theme.spacing(1),
+            marginRight: theme.spacing(1),
             width: 200,
         },
         textField2: {
-            marginLeft: theme.spacing.unit,
-            marginRight: theme.spacing.unit,
+            marginLeft: theme.spacing(1),
+            marginRight: theme.spacing(1),
             width: 400,
         },
         button: {
-            margin: 0.1 * theme.spacing.unit,
+            margin: 0.1 * theme.spacing(1),
         },
     });
 
@@ -215,7 +215,7 @@ class ServerDialog extends React.Component<ServerDialogProps, ServerDialogState>
         let dialogContent;
         if (dialogMode === "add" || dialogMode === "edit") {
             dialogContent = (
-                <DialogContent>
+                <DialogContent dividers>
                     <TextField
                         required
                         id="server-name"
@@ -239,7 +239,7 @@ class ServerDialog extends React.Component<ServerDialogProps, ServerDialogState>
             );
         } else {
             dialogContent = (
-                <DialogContent>
+                <DialogContent dividers>
                     <div>
                         <FormControl className={classes.formControl}>
                             <InputLabel htmlFor="server-name">Name</InputLabel>
