@@ -5,9 +5,7 @@ import { Theme } from '@material-ui/core';
 import { WithLocale } from '../util/lang';
 import TimeSeriesChart from './TimeSeriesChart';
 import { Time, TimeRange, TimeSeriesGroup } from '../model/timeSeries';
-// import Card from '@material-ui/core/Card';
-// import CardContent from '@material-ui/core/CardContent';
-// import Typography from '@material-ui/core/Typography';
+import { PlaceInfo } from "../model/place";
 import TimeRangeSlider from './TimeRangeSlider';
 
 
@@ -36,13 +34,15 @@ interface TimeSeriesChartsProps extends WithStyles<typeof styles>, WithLocale {
     selectedTime?: Time | null;
     selectTime?: (time: Time | null) => void;
 
+    showPointsOnly: boolean;
+    showErrorBars: boolean;
+
     dataTimeRange?: TimeRange | null;
     selectedTimeRange?: TimeRange | null;
     selectTimeRange?: (timeRange: TimeRange | null) => void;
-    visibleTimeRange?: TimeRange | null;
-    updateVisibleTimeRange?: (timeRange: TimeRange | null) => void;
 
     removeTimeSeriesGroup?: (id: string) => void;
+    placeInfos?: { [placeId: string]: PlaceInfo };
 }
 
 class TimeSeriesCharts extends React.Component<TimeSeriesChartsProps> {
@@ -53,11 +53,9 @@ class TimeSeriesCharts extends React.Component<TimeSeriesChartsProps> {
             timeSeriesGroups,
             selectedTime, selectedTimeRange,
             dataTimeRange, selectTime, selectTimeRange,
-            removeTimeSeriesGroup,
-            visibleTimeRange,
-            updateVisibleTimeRange,
-        }
-            = this.props;
+            removeTimeSeriesGroup, showPointsOnly, showErrorBars,
+            placeInfos
+        } = this.props;
         const charts = timeSeriesGroups.map(timeSeriesGroup => (
             <TimeSeriesChart
                 key={timeSeriesGroup.id}
@@ -69,6 +67,9 @@ class TimeSeriesCharts extends React.Component<TimeSeriesChartsProps> {
                 selectTime={selectTime}
                 selectTimeRange={selectTimeRange}
                 removeTimeSeriesGroup={removeTimeSeriesGroup}
+                showPointsOnly={showPointsOnly}
+                showErrorBars={showErrorBars}
+                placeInfos={placeInfos}
             />)
         );
         if (charts.length > 0) {
@@ -78,8 +79,6 @@ class TimeSeriesCharts extends React.Component<TimeSeriesChartsProps> {
                         selectedTimeRange={selectedTimeRange}
                         dataTimeRange={dataTimeRange}
                         selectTimeRange={selectTimeRange}
-                        visibleTimeRange={visibleTimeRange}
-                        updateVisibleTimeRange={updateVisibleTimeRange}
                     />
                     {charts}
                 </div>
