@@ -2,10 +2,11 @@ import { connect } from 'react-redux';
 
 import PlaceSelect from "../components/PlaceSelect";
 import { AppState } from '../states/appState';
+import { removeUserPlace } from "../actions/dataActions";
 import { selectPlace } from '../actions/controlActions';
 import {
-    selectedDatasetSelectedPlaceGroupPlacesSelector,
-    selectedDatasetSelectedPlaceGroupPlaceLabelsSelector
+    selectedPlaceGroupPlacesSelector,
+    selectedPlaceGroupPlaceLabelsSelector
 } from '../selectors/controlSelectors';
 
 
@@ -14,15 +15,18 @@ const mapStateToProps = (state: AppState) => {
         locale: state.controlState.locale,
 
         datasets: state.dataState.datasets,
+        userPlaceGroup: state.dataState.userPlaceGroup,
 
+        selectedPlaceGroupIds: state.controlState.selectedPlaceGroupIds,
         selectedPlaceId: state.controlState.selectedPlaceId,
-        places: selectedDatasetSelectedPlaceGroupPlacesSelector(state),
-        placeLabels: selectedDatasetSelectedPlaceGroupPlaceLabelsSelector(state),
+        places: selectedPlaceGroupPlacesSelector(state),
+        placeLabels: selectedPlaceGroupPlaceLabelsSelector(state),
     };
 };
 
 const mapDispatchToProps = {
     selectPlace,
+    removeUserPlace,
 };
 
 export default connect(mapStateToProps, mapDispatchToProps)(PlaceSelect);
