@@ -2,7 +2,7 @@ import { connect } from 'react-redux';
 
 import { AppState } from '../states/appState';
 import TimeSelect from '../components/TimeSelect';
-import { selectTime, selectTimeRange, incSelectedTime, updateTimeAnimation } from '../actions/controlActions';
+import { selectTime } from '../actions/controlActions';
 
 
 const mapStateToProps = (state: AppState) => {
@@ -11,16 +11,11 @@ const mapStateToProps = (state: AppState) => {
 
         selectedTime: state.controlState.selectedTime,
         selectedTimeRange: state.controlState.selectedTimeRange,
-        timeAnimationActive: state.controlState.timeAnimationActive,
-        timeAnimationInterval: state.controlState.timeAnimationInterval,
     };
 };
 
 const mapDispatchToProps = {
     selectTime,
-    selectTimeRange,
-    incSelectedTime,
-    updateTimeAnimation,
 };
 
 export default connect(mapStateToProps, mapDispatchToProps)(TimeSelect);
