@@ -4,7 +4,7 @@ import { AppState } from '../states/appState';
 import TimeSeriesCharts from '../components/TimeSeriesCharts';
 import { removeTimeSeriesGroup } from "../actions/dataActions";
 import { selectTime, selectTimeRange } from "../actions/controlActions";
-import { selectedDatasetTimeRangeSelector } from '../selectors/controlSelectors';
+import { selectedDatasetTimeRangeSelector, timeSeriesPlaceInfosSelector } from '../selectors/controlSelectors';
 
 
 const mapStateToProps = (state: AppState) => {
@@ -14,6 +14,9 @@ const mapStateToProps = (state: AppState) => {
         selectedTime: state.controlState.selectedTime,
         selectedTimeRange: state.controlState.selectedTimeRange,
         dataTimeRange: selectedDatasetTimeRangeSelector(state),
+        showPointsOnly: state.controlState.showTimeSeriesPointsOnly,
+        showErrorBars: state.controlState.showTimeSeriesErrorBars,
+        placeInfos: timeSeriesPlaceInfosSelector(state),
     }
 };
 
