@@ -2,7 +2,7 @@ import { connect } from 'react-redux';
 
 import { AppState } from '../states/appState';
 import TimeSlider from '../components/TimeSlider';
-import { selectTime, selectTimeRange, updateVisibleTimeRange } from '../actions/controlActions';
+import { selectTime, selectTimeRange } from '../actions/controlActions';
 
 
 const mapStateToProps = (state: AppState) => {
@@ -11,14 +11,12 @@ const mapStateToProps = (state: AppState) => {
 
         selectedTime: state.controlState.selectedTime,
         selectedTimeRange: state.controlState.selectedTimeRange,
-        visibleTimeRange: state.controlState.visibleTimeRange,
     };
 };
 
 const mapDispatchToProps = {
     selectTime,
     selectTimeRange,
-    updateVisibleTimeRange,
 };
 
 export default connect(mapStateToProps, mapDispatchToProps)(TimeSlider);
