@@ -2,8 +2,9 @@ import { connect } from 'react-redux';
 
 import VariableSelect from "../components/VariableSelect";
 import { AppState } from '../states/appState';
+import { addTimeSeries } from "../actions/dataActions";
 import { selectVariable } from '../actions/controlActions';
-import { selectedDatasetVariablesSelector } from '../selectors/controlSelectors';
+import { canAddTimeSeriesSelector, selectedDatasetVariablesSelector } from '../selectors/controlSelectors';
 
 
 const mapStateToProps = (state: AppState) => {
@@ -11,12 +12,14 @@ const mapStateToProps = (state: AppState) => {
         locale: state.controlState.locale,
 
         selectedVariableName: state.controlState.selectedVariableName,
+        canAddTimeSeries: canAddTimeSeriesSelector(state),
         variables: selectedDatasetVariablesSelector(state),
     };
 };
 
 const mapDispatchToProps = {
     selectVariable,
+    addTimeSeries,
 };
 
 export default connect(mapStateToProps, mapDispatchToProps)(VariableSelect);
