@@ -6,8 +6,9 @@ import {
     selectedDatasetPlaceGroupLayersSelector,
     selectedDatasetVariableLayerSelector
 } from '../selectors/controlSelectors';
-import { addGeometry } from '../actions/controlActions';
+import { addUserPlace } from '../actions/dataActions';
 import Viewer from '../components/Viewer';
+import { userPlaceGroupSelector } from "../selectors/dataSelectors";
 
 
 const mapStateToProps = (state: AppState) => {
@@ -16,13 +17,14 @@ const mapStateToProps = (state: AppState) => {
         variableLayer: selectedDatasetVariableLayerSelector(state),
         placeGroupLayers: selectedDatasetPlaceGroupLayersSelector(state),
         colorBarLegend: selectedColorBarLegendSelector(state),
+        userPlaceGroup: userPlaceGroupSelector(state),
         drawMode: state.controlState.selectedDrawMode,
         flyTo: state.controlState.flyTo,
     }
 };
 
 const mapDispatchToProps = {
-    addGeometry,
+    addUserPlace,
 };
 
 export default connect(mapStateToProps, mapDispatchToProps)(Viewer);
