@@ -1,7 +1,6 @@
 import * as ol from 'openlayers';
 
-import { findDataset, findDatasetVariable, findDatasetPlace, getDatasetTimeRange } from '../model/dataset';
-import { TimeRange } from "../model/timeSeries";
+import { findDataset, findDatasetVariable, getDatasetTimeRange, findDatasetOrUserPlace } from '../model/dataset';
 import { ControlState, newControlState } from '../states/controlState';
 import {
     SELECT_DATASET,
@@ -10,22 +9,23 @@ import {
     SELECT_PLACE,
     SELECT_TIME,
     SELECT_TIME_SERIES_UPDATE_MODE,
-    SELECT_USER_PLACE,
-    ADD_GEOMETRY,
-    SELECT_GEOMETRY,
     ControlAction,
     SELECT_TIME_RANGE,
-    UPDATE_VISIBLE_TIME_RANGE,
     UPDATE_TIME_ANIMATION,
     ADD_ACTIVITY,
     REMOVE_ACTIVITY,
-    CHANGE_LOCALE, OPEN_DIALOG, CLOSE_DIALOG, INC_SELECTED_TIME,
+    CHANGE_LOCALE,
+    OPEN_DIALOG,
+    CLOSE_DIALOG,
+    INC_SELECTED_TIME,
+    UPDATE_SETTINGS,
 } from '../actions/controlActions';
-import { CONFIGURE_SERVERS, DataAction } from "../actions/dataActions";
+import { CONFIGURE_SERVERS, DataAction, ADD_USER_PLACE } from "../actions/dataActions";
 import { I18N } from "../config";
 import { AppState } from "../states/appState";
 import { selectedTimeIndexSelector, timeCoordinatesSelector } from "../selectors/controlSelectors";
 import { findIndexCloseTo } from "../util/find";
+import { storeUserSettings } from '../states/userSettings';
 
 
 const SIMPLE_GEOMETRY_TYPES = ['Point', 'LineString', 'LinearRing', 'Polygon', 'MultiPoint', 'MultiLineString', 'MultiPolygon', 'Circle'];
@@ -35,6 +35,9 @@ export function controlReducer(state: ControlState, action: ControlAction | Data
         state = newControlState();
     }
     switch (action.type) {
+        case UPDATE_SETTINGS:
+            storeUserSettings(action.settings);
+            return action.settings;
         case SELECT_DATASET: {
             let selectedVariableName = state.selectedVariableName;
             const dataset = findDataset(action.datasets, action.selectedDatasetId)!;
@@ -48,31 +51,30 @@ export function controlReducer(state: ControlState, action: ControlAction | Data
             }
             const selectedDatasetId = action.selectedDatasetId;
             const selectedTimeRange = getDatasetTimeRange(dataset);
-            const visibleTimeRange: TimeRange | null = selectedTimeRange ? [selectedTimeRange[0], selectedTimeRange[1]] : null;
             const selectedTime = selectedTimeRange ? selectedTimeRange[1] : null;
             return {
                 ...state,
                 selectedDatasetId,
                 selectedVariableName,
                 selectedTimeRange,
-                visibleTimeRange,
                 selectedTime,
                 flyTo,
             };
         }
         case SELECT_PLACE_GROUPS: {
             const selectedPlaceGroupIds = action.selectedPlaceGroupIds;
+            const selectedPlaceId = null;
             return {
                 ...state,
                 selectedPlaceGroupIds,
+                selectedPlaceId,
             };
         }
         case SELECT_PLACE: {
             const selectedPlaceId = action.selectedPlaceId;
             let flyTo = state.flyTo;
             if (selectedPlaceId) {
-                const dataset = findDataset(action.datasets, state.selectedDatasetId)!;
-                const place = findDatasetPlace(dataset, selectedPlaceId);
+                const place = findDatasetOrUserPlace(action.datasets, action.userPlaceGroup, selectedPlaceId);
                 if (place !== null) {
                     if (place.bbox && place.bbox.length === 4) {
                         flyTo = place.bbox as [number, number, number, number];
@@ -91,12 +93,6 @@ export function controlReducer(state: ControlState, action: ControlAction | Data
             return {
                 ...state,
                 selectedVariableName: action.selectedVariableName,
-            };
-        }
-        case SELECT_USER_PLACE: {
-            return {
-                ...state,
-                selectedUserPlaceId: action.selectedUserPlaceId,
             };
         }
         case SELECT_TIME: {
@@ -152,12 +148,6 @@ export function controlReducer(state: ControlState, action: ControlAction | Data
                 selectedTimeRange: action.selectedTimeRange,
             };
         }
-        case UPDATE_VISIBLE_TIME_RANGE: {
-            return {
-                ...state,
-                visibleTimeRange: action.visibleTimeRange,
-            };
-        }
         case SELECT_TIME_SERIES_UPDATE_MODE: {
             return {
                 ...state,
@@ -171,16 +161,19 @@ export function controlReducer(state: ControlState, action: ControlAction | Data
                 timeAnimationInterval: action.timeAnimationInterval,
             };
         }
-        case ADD_GEOMETRY: {
+        case ADD_USER_PLACE: {
+            let selectedPlaceGroupIds;
+            if (!state.selectedPlaceGroupIds || state.selectedPlaceGroupIds.length === 0) {
+                selectedPlaceGroupIds = ['user'];
+            } else if (state.selectedPlaceGroupIds.find(id => id === 'user')) {
+                selectedPlaceGroupIds = state.selectedPlaceGroupIds;
+            } else {
+                selectedPlaceGroupIds = [...state.selectedPlaceGroupIds, 'user']
+            }
             return {
                 ...state,
-                selectedFeatureId: action.featureId,
-            };
-        }
-        case SELECT_GEOMETRY: {
-            return {
-                ...state,
-                selectedFeatureId: action.featureId,
+                selectedPlaceGroupIds,
+                selectedPlaceId: action.id,
             };
         }
         case ADD_ACTIVITY: {
