@@ -1,17 +1,20 @@
 import * as ol from 'openlayers';
 
-import { DataState, newDataState, storeUserServers } from '../states/dataState';
+import { DataState, newDataState } from '../states/dataState';
+import { storeUserServers } from '../states/userSettings';
 import {
     CONFIGURE_SERVERS,
+    ADD_USER_PLACE,
     DataAction, REMOVE_ALL_TIME_SERIES,
     REMOVE_TIME_SERIES_GROUP,
     UPDATE_COLOR_BARS,
     UPDATE_DATASETS,
-    UPDATE_TIME_SERIES, UPDATE_DATASET_PLACE_GROUP
+    UPDATE_TIME_SERIES, UPDATE_DATASET_PLACE_GROUP, REMOVE_USER_PLACE, REMOVE_ALL_USER_PLACES
 } from '../actions/dataActions';
 import { MAP_OBJECTS } from '../states/controlState';
 import { newId } from '../util/id';
-import { TimeSeriesGroup } from '../model/timeSeries';
+import { Place } from "../model/place";
+import { TimeSeries, TimeSeriesGroup } from "../model/timeSeries";
 
 
 export function dataReducer(state: DataState, action: DataAction): DataState {
@@ -39,87 +42,82 @@ export function dataReducer(state: DataState, action: DataAction): DataState {
             }
             return state;
         }
+        case ADD_USER_PLACE: {
+            const {id, label, color, geometry} = action;
+            const feature = {
+                type: 'Feature',
+                id,
+                geometry,
+                properties: {label, color},
+            };
+            const features = [...state.userPlaceGroup.features, feature as Place];
+            const userPlaceGroup = {...state.userPlaceGroup, features};
+            return {
+                ...state,
+                userPlaceGroup,
+            };
+        }
+        case REMOVE_USER_PLACE: {
+            const {id} = action;
+            const index = state.userPlaceGroup.features.findIndex(p => p.id == id);
+            if (index >= 0) {
+                removeUserPlacesFromLayer([id]);
+                const features = [...state.userPlaceGroup.features];
+                features.splice(index, 1);
+                const userPlaceGroup = {...state.userPlaceGroup, features};
+                const timeSeriesArray = getTimeSeriesArray(state.timeSeriesGroups, [id]);
+                let timeSeriesGroups = state.timeSeriesGroups;
+                timeSeriesArray.forEach(ts => {
+                    timeSeriesGroups = updateTimeSeriesGroups(timeSeriesGroups,
+                                                              ts, 'remove', 'append');
+                });
+                return {
+                    ...state,
+                    userPlaceGroup,
+                    timeSeriesGroups
+                };
+            }
+            return state;
+        }
+        case REMOVE_ALL_USER_PLACES: {
+            const userPlaces = state.userPlaceGroup.features as Place[];
+            const userPlaceIds = userPlaces.map(p => p.id);
+            removeUserPlacesFromLayer(userPlaceIds);
+            const userPlaceGroup = {...state.userPlaceGroup, features: []};
+            const timeSeriesArray = getTimeSeriesArray(state.timeSeriesGroups, userPlaceIds);
+            let timeSeriesGroups = state.timeSeriesGroups;
+            timeSeriesArray.forEach(ts => {
+                timeSeriesGroups = updateTimeSeriesGroups(timeSeriesGroups,
+                                                          ts, 'remove', 'append');
+            });
+            return {
+                ...state,
+                userPlaceGroup,
+                timeSeriesGroups,
+            };
+        }
         case UPDATE_COLOR_BARS: {
             return {...state, colorBars: action.colorBars};
         }
         case UPDATE_TIME_SERIES: {
-            let newTimeSeries = action.timeSeries;
-
-            if (action.updateMode === 'replace' && action.dataMode === 'new') {
-                const featureId = newTimeSeries.source.featureId;
-                state.timeSeriesGroups.forEach(tsg => {
-                    removeTimeSeriesGroupFeatures(tsg, (fid: string) => fid !== featureId);
-                });
-            }
-
-            let newTimeSeriesGroups;
-            const tsgIndex = state.timeSeriesGroups.findIndex(tsg => tsg.variableUnits === newTimeSeries.source.variableUnits);
-            if (tsgIndex >= 0) {
-                const timeSeriesGroup = state.timeSeriesGroups[tsgIndex];
-                const timeSeriesArray = timeSeriesGroup.timeSeriesArray;
-                const tsIndex = timeSeriesArray.findIndex(ts => ts.source.datasetId === newTimeSeries.source.datasetId
-                                                                && ts.source.featureId === newTimeSeries.source.featureId
-                                                                && ts.source.variableName === newTimeSeries.source.variableName);
-                let newTimeSeriesArray;
-                if (tsIndex >= 0) {
-                    const oldTimeSeries = timeSeriesArray[tsIndex];
-                    if (action.dataMode === 'append') {
-                        newTimeSeries = {...newTimeSeries, data: [...newTimeSeries.data, ...oldTimeSeries.data]};
-                    }
-                    if (action.updateMode === 'replace') {
-                        newTimeSeriesArray = [newTimeSeries];
-                    } else {
-                        newTimeSeriesArray = timeSeriesArray.slice();
-                        newTimeSeriesArray[tsIndex] = newTimeSeries;
-                    }
-                } else {
-                    if (action.updateMode === 'replace') {
-                        newTimeSeriesArray = [newTimeSeries];
-                    } else {
-                        newTimeSeriesArray = [newTimeSeries, ...timeSeriesArray];
-                    }
-                }
-                const newTimeSeriesGroup = {...timeSeriesGroup, timeSeriesArray: newTimeSeriesArray};
-                if (action.updateMode === 'replace') {
-                    newTimeSeriesGroups = [newTimeSeriesGroup];
-                } else {
-                    newTimeSeriesGroups = state.timeSeriesGroups.slice();
-                    newTimeSeriesGroups[tsgIndex] = newTimeSeriesGroup;
-                }
-            } else {
-                const newTimeSeriesGroup = {
-                    id: newId(),
-                    variableUnits: newTimeSeries.source.variableUnits,
-                    timeSeriesArray: [newTimeSeries],
-                };
-                if (action.updateMode === 'replace') {
-                    newTimeSeriesGroups = [newTimeSeriesGroup];
-                } else {
-                    newTimeSeriesGroups = [newTimeSeriesGroup, ...state.timeSeriesGroups];
-                }
-            }
-            if (newTimeSeriesGroups) {
-                return {
-                    ...state,
-                    timeSeriesGroups: newTimeSeriesGroups,
-                };
+            const {timeSeries, updateMode, dataMode} = action;
+            const timeSeriesGroups = updateTimeSeriesGroups(state.timeSeriesGroups,
+                                                            timeSeries, updateMode, dataMode);
+            if (timeSeriesGroups !== state.timeSeriesGroups) {
+                return {...state, timeSeriesGroups};
             }
             return state;
         }
         case REMOVE_TIME_SERIES_GROUP: {
             let tsgIndex = state.timeSeriesGroups.findIndex(tsg => tsg.id === action.id);
             if (tsgIndex >= 0) {
-                removeTimeSeriesGroupFeatures(state.timeSeriesGroups[tsgIndex]);
-                const timeSeriesGroups = state.timeSeriesGroups.slice();
+                const timeSeriesGroups = [...state.timeSeriesGroups];
                 timeSeriesGroups.splice(tsgIndex, 1);
                 return {...state, timeSeriesGroups};
             }
             return state;
         }
         case REMOVE_ALL_TIME_SERIES: {
-            state.timeSeriesGroups.forEach(tsg => {
-                removeTimeSeriesGroupFeatures(tsg);
-            });
             return {...state, timeSeriesGroups: []};
         }
         case CONFIGURE_SERVERS: {
@@ -135,18 +133,111 @@ export function dataReducer(state: DataState, action: DataAction): DataState {
 }
 
 
-function removeTimeSeriesGroupFeatures(tsg: TimeSeriesGroup, predicate?: (fid: string) => boolean) {
+function removeUserPlacesFromLayer(userPlaceIds: string[]) {
     if (MAP_OBJECTS.userLayer) {
         const userLayer = MAP_OBJECTS.userLayer as ol.layer.Vector;
-        tsg.timeSeriesArray.forEach(ts => {
-            const fid = ts.source.featureId;
-            if (!predicate || predicate(fid)) {
-                const source = userLayer.getSource();
-                const feature = source.getFeatureById(fid);
-                if (feature) {
-                    source.removeFeature(feature);
-                }
+        const source = userLayer.getSource();
+        userPlaceIds.forEach(placeId => {
+            const feature = source.getFeatureById(placeId);
+            if (feature) {
+                source.removeFeature(feature);
             }
         });
     }
+}
+
+
+function updateTimeSeriesGroups(timeSeriesGroups: TimeSeriesGroup[],
+                                timeSeries: TimeSeries,
+                                updateMode: 'add' | 'replace' | 'remove',
+                                dataMode: 'append' | 'new'): TimeSeriesGroup[] {
+    let currentTimeSeries = timeSeries;
+
+    let newTimeSeriesGroups;
+    const tsgIndex = timeSeriesGroups.findIndex(tsg => tsg.variableUnits === currentTimeSeries.source.variableUnits);
+    if (tsgIndex >= 0) {
+        const timeSeriesGroup = timeSeriesGroups[tsgIndex];
+        const timeSeriesArray = timeSeriesGroup.timeSeriesArray;
+        const tsIndex = timeSeriesArray.findIndex(ts => ts.source.datasetId === currentTimeSeries.source.datasetId
+                                                        && ts.source.variableName === currentTimeSeries.source.variableName
+                                                        && ts.source.placeId === currentTimeSeries.source.placeId);
+        let newTimeSeriesArray;
+        if (tsIndex >= 0) {
+            const oldTimeSeries = timeSeriesArray[tsIndex];
+            if (dataMode === 'append') {
+                currentTimeSeries = {
+                    ...currentTimeSeries,
+                    data: [...currentTimeSeries.data, ...oldTimeSeries.data]
+                };
+            }
+            if (updateMode === 'replace') {
+                newTimeSeriesArray = [currentTimeSeries];
+            } else if (updateMode === 'add') {
+                newTimeSeriesArray = timeSeriesArray.slice();
+                newTimeSeriesArray[tsIndex] = currentTimeSeries;
+            } else /*if (action.updateMode === 'remove')*/ {
+                newTimeSeriesArray = timeSeriesArray.slice();
+                newTimeSeriesArray.splice(tsIndex, 1);
+            }
+        } else {
+            if (updateMode === 'replace') {
+                newTimeSeriesArray = [currentTimeSeries];
+            } else if (updateMode === 'add') {
+                newTimeSeriesArray = [currentTimeSeries, ...timeSeriesArray];
+            } else /*if (action.updateMode === 'remove')*/ {
+                // Nothing to do here.
+                newTimeSeriesArray = timeSeriesArray;
+            }
+        }
+        if (updateMode === 'replace') {
+            newTimeSeriesGroups = [{...timeSeriesGroup, timeSeriesArray: newTimeSeriesArray}];
+        } else if (updateMode === 'add') {
+            newTimeSeriesGroups = timeSeriesGroups.slice();
+            newTimeSeriesGroups[tsgIndex] = {...timeSeriesGroup, timeSeriesArray: newTimeSeriesArray};
+        } else /*if (action.updateMode === 'remove')*/ {
+            if (newTimeSeriesArray.length >= 0) {
+                newTimeSeriesGroups = timeSeriesGroups.slice();
+                newTimeSeriesGroups[tsgIndex] = {...timeSeriesGroup, timeSeriesArray: newTimeSeriesArray};
+            } else {
+                newTimeSeriesGroups = timeSeriesGroups.slice();
+                newTimeSeriesGroups.splice(tsgIndex, 1)
+            }
+        }
+    } else {
+        if (updateMode === 'replace') {
+            const newTimeSeriesGroup = {
+                id: newId(),
+                variableUnits: currentTimeSeries.source.variableUnits,
+                timeSeriesArray: [currentTimeSeries],
+            };
+            newTimeSeriesGroups = [newTimeSeriesGroup];
+        } else if (updateMode === 'add') {
+            const newTimeSeriesGroup = {
+                id: newId(),
+                variableUnits: currentTimeSeries.source.variableUnits,
+                timeSeriesArray: [currentTimeSeries],
+            };
+            newTimeSeriesGroups = [newTimeSeriesGroup, ...timeSeriesGroups];
+        } else /*if (action.updateMode === 'remove')*/ {
+            // Nothing to do here.
+            newTimeSeriesGroups = timeSeriesGroups;
+        }
+    }
+
+    return newTimeSeriesGroups;
+}
+
+
+function getTimeSeriesArray(timeSeriesGroups: TimeSeriesGroup[], placeIds: string[]): TimeSeries[] {
+    const timeSeriesArray: TimeSeries[] = [];
+    timeSeriesGroups.forEach(tsg => {
+        tsg.timeSeriesArray.forEach(ts => {
+            placeIds.forEach(placeId => {
+                if (ts.source.placeId === placeId) {
+                    timeSeriesArray.push(ts);
+                }
+            });
+        })
+    });
+    return timeSeriesArray;
 }
