@@ -1,17 +1,17 @@
-import { Dataset, } from '../model/dataset';
+import { Dataset } from '../model/dataset';
 import { Place, PlaceGroup, } from '../model/place';
 import { TimeSeriesGroup } from '../model/timeSeries';
 import { ColorBars } from '../model/colorBar';
 import { Server } from "../model/server";
-import { VIEWER_API_SERVERS } from "../config";
-import { getLocalStorage } from "../util/storage";
+import { I18N, VIEWER_API_SERVERS } from "../config";
+import { loadUserServers } from './userSettings';
 
 
 export interface DataState {
     datasets: Dataset[];
     colorBars: ColorBars | null;
     timeSeriesGroups: TimeSeriesGroup[];
-    userPlaces: PlaceGroup;
+    userPlaceGroup: PlaceGroup;
     userServers: Server[];
 }
 
@@ -27,38 +27,13 @@ export function newDataState(): DataState {
         datasets: [],
         colorBars: null,
         timeSeriesGroups: [],
-        userPlaces: {id: 'user', title: 'My places', type: "FeatureCollection", features: [] as Array<Place>},
+        userPlaceGroup: {
+            id: 'user',
+            title: I18N.get('My places'),
+            type: "FeatureCollection",
+            features: [] as Array<Place>
+        },
         userServers,
     };
 }
-
-export function storeUserServers(userServers: Server[]) {
-    const storage = getLocalStorage();
-    if (storage) {
-        let userServersJson = JSON.stringify(userServers);
-        try {
-            storage.setItem("xcube.userServers", userServersJson);
-        } catch (e) {
-            console.warn(`failed to store user servers: ${e}`);
-        }
-    }
-}
-
-export function loadUserServers(): Server[] {
-    const storage = getLocalStorage();
-    if (storage) {
-        let userServersJson;
-        try {
-            userServersJson = storage.getItem("xcube.userServers");
-        } catch (e) {
-            console.warn(`failed to store user servers: ${e}`);
-        }
-        if (userServersJson) {
-            return JSON.parse(userServersJson) as Server[];
-        }
-    }
-    return [];
-}
-
-
 
